backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic[email]==2.9.2
pydantic-settings==2.5.2
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
redis==5.0.7
prometheus-fastapi-instrumentator==7.0.0

pytest 
pytest-asyncio 
httpx